// apps/api/src/core/backends/_enum/action-type.enum.ts
export enum ActionType {
  LIST_BACKENDS = 'LIST_BACKENDS',
  DUMP_PACKAGE_CONFIG = 'DUMP_PACKAGE_CONFIG',
  PING_TARGET = 'PING_TARGET',
  IDENTITY_CREATE = 'IDENTITY_CREATE',
  IDENTITY_UPDATE = 'IDENTITY_UPDATE',
  IDENTITY_DELETE = 'IDENTITY_DELETE',
  IDENTITY_PASSWORD_RESET = 'IDENTITY_PASSWORD_RESET',
  IDENTITY_PASSWORD_CHANGE = 'IDENTITY_PASSWORD_CHANGE',
  IDENTITY_DISABLE = 'IDENTITY_DISABLE',
  IDENTITY_ENABLE = 'IDENTITY_ENABLE',
}